Add high-level eta.core requirements to setup.py
Per https://packaging.python.org/discussions/install-requires-vs-requirements/, only direct dependencies are listed with no specific versions. OpenCV and Pillow are restricted to the current major version as they tend to break compatibility in new major releases.

Tensorflow was skipped due to it not being required for a lightweight install. Some detectors rely on it, as well as the following modules:
- label_map_util
- nets
- preprocessing

The following packages were skipped due to being development-specific:
- docutils
- sphinx

The following packages were skipped because nothing in ETA appears to depend on them directly - unsure why they were in requirements.txt:
- blockdiag
- contextlib2
- Cython
- funcparserlib
- Jinja2
- lxml
- MarkupSafe
- oauth2client
- requests-toolbelt
- scikit-learn
- scipy
- simplejson
- webcolors